Conflicts:
	lib/djgpp.djl
	src/libc/ansi/stdio/doscan.c
	src/libc/crt0/crt0.S
	src/libc/crt0/crt1.c
	src/makefile.inc
	src/makefile.lib
	src/stub/makefile

+++ include/pc.h
@@ -11,6 +11,8 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
+
 #ifndef __dj_ENFORCE_ANSI_FREESTANDING
 
 #if (defined(__STDC_VERSION__) && __STDC_VERSION__ >= 199901L) \
@@ -24,16 +26,16 @@ extern "C" {
 
 unsigned char	inportb  (unsigned short _port);
 unsigned short	inportw  (unsigned short _port);
-unsigned long	inportl  (unsigned short _port);
+uint32_t	inportl  (unsigned short _port);
 void		inportsb (unsigned short _port, unsigned char  *_buf, unsigned _len);
 void		inportsw (unsigned short _port, unsigned short *_buf, unsigned _len);
-void		inportsl (unsigned short _port, unsigned long  *_buf, unsigned _len);
+void		inportsl (unsigned short _port, uint32_t      *_buf, unsigned _len);
 void		outportb (unsigned short _port, unsigned char  _data);
 void		outportw (unsigned short _port, unsigned short _data);
-void		outportl (unsigned short _port, unsigned long  _data);
+void		outportl (unsigned short _port, uint32_t _data);
 void		outportsb(unsigned short _port, const unsigned char  *_buf, unsigned _len);
 void		outportsw(unsigned short _port, const unsigned short *_buf, unsigned _len);
-void		outportsl(unsigned short _port, const unsigned long  *_buf, unsigned _len);
+void		outportsl(unsigned short _port, const uint32_t *_buf, unsigned _len);
 
 unsigned char	inp(unsigned short _port);
 unsigned short	inpw(unsigned short _port);
