+++ src/libemu/src/emu387.cc
@@ -428,7 +428,7 @@ static void *get_modrm()
 {
   int mod = modrm>>6;
   int rm = modrm & 7;
-  int rv = 0;
+  unsigned long rv = 0;
   switch (mod)
   {
     case 0:
